refactor: extract grading orchestration to simulation/grading.py (Architect Rec 2)
Move _assign_outcomes, _assign_outcomes_absolute, _assign_outcomes_relative,
_filter_eligible_states, _classify_absolute_single, _classify_relative, and
_compute_midterm_aggregate from SimulationEngine to free functions in grading.py.
Public entry point is assign_outcomes(states, grading_config, engine_config, rng).
engine.py drops ~170 lines (760 -> 590). Tests updated to call free function directly.
